@@ -1,18 +1,9 @@
 {
   "name": "indoor-gas-network",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Розрахунок внутрышньо-будинкової газової мережі",
-  "main": "gulpfile.js",
-  "scripts": {
-    "start": "gulp",
-    "build": "cross-env NODE_ENV=production gulp build",
-    "deploy": "npm run build && gh-pages -d build",
-    "dist": "npm run build && gulp zip",
-    "lintspaces": "gulp lintspaces",
-    "stylelint": "stylelint \"source/sass/**/*.{scss,sass}\" --syntax scss",
-    "eslint": "eslint \"source/js/**/*.js\"",
-    "test": "npm run lintspaces && npm run stylelint && npm run eslint"
-  },
+  "type": "module",
+  "private": true,
   "repository": {
     "type": "git",
     "url": "git+ssh://git@github.com/embyth/indoor-gas-network.git"
@@ -23,45 +14,58 @@
     "url": "https://github.com/embyth/indoor-gas-network/issues"
   },
   "homepage": "https://github.com/embyth/indoor-gas-network#readme",
+  "scripts": {
+    "start": "gulp",
+    "build": "cross-env NODE_ENV=production gulp build",
+    "deploy": "npm run build && gh-pages -d build",
+    "dist": "npm run build && gulp zip",
+    "convertfonts": "gulp convertfonts",
+    "webp": "gulp webp",
+    "sprite": "gulp sprite",
+    "lintspaces": "gulp lintspaces",
+    "stylelint": "stylelint \"source/sass/**/*.{scss,sass}\" --syntax scss",
+    "eslint": "eslint \"source/js/**/*.js\"",
+    "test": "npm run lintspaces && npm run stylelint && npm run eslint"
+  },
   "devDependencies": {
-    "@babel/core": "^7.11.1",
-    "@babel/preset-env": "^7.11.0",
-    "autoprefixer": "^9.8.6",
-    "browser-sync": "^2.26.12",
-    "cross-env": "^7.0.2",
-    "del": "^5.1.0",
-    "eslint": "^7.6.0",
-    "gh-pages": "^3.1.0",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-changed": "^4.0.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-csso": "^4.0.1",
-    "gulp-htmlmin": "^5.0.1",
-    "gulp-if": "^3.0.0",
-    "gulp-imagemin": "^7.1.0",
-    "gulp-lintspaces": "^1.0.1",
-    "gulp-plumber": "^1.2.1",
-    "gulp-postcss": "^8.0.0",
-    "gulp-rename": "^2.0.0",
-    "gulp-sass": "^4.1.0",
-    "gulp-svgstore": "^7.0.1",
-    "gulp-ttf2woff": "^1.1.1",
-    "gulp-ttf2woff2": "^3.0.0",
-    "gulp-uglify": "^3.0.2",
-    "gulp-webp": "^4.0.1",
-    "gulp-zip": "^5.0.2",
-    "husky": "^4.2.5",
-    "imagemin-jpegoptim": "^7.0.0",
-    "require-dir": "^1.2.0",
-    "stylelint": "^13.6.1",
-    "stylelint-config-standard": "^20.0.0"
+    "@babel/core": "7.12.8",
+    "@babel/preset-env": "7.12.7",
+    "autoprefixer": "9.8.6",
+    "babel-loader": "8.2.1",
+    "browser-sync": "2.26.13",
+    "cross-env": "7.0.2",
+    "del": "6.0.0",
+    "eslint": "7.14.0",
+    "eslint-config-htmlacademy": "0.6.0",
+    "gh-pages": "3.1.0",
+    "gulp": "4.0.2",
+    "gulp-changed": "4.0.2",
+    "gulp-csso": "4.0.1",
+    "gulp-htmlmin": "5.0.1",
+    "gulp-if": "3.0.0",
+    "gulp-imagemin": "7.1.0",
+    "gulp-lintspaces": "1.0.1",
+    "gulp-plumber": "1.2.1",
+    "gulp-postcss": "9.0.0",
+    "gulp-rename": "2.0.0",
+    "gulp-sass": "4.1.0",
+    "gulp-svgstore": "7.0.1",
+    "gulp-ttf2woff": "1.1.1",
+    "gulp-ttf2woff2": "3.0.0",
+    "gulp-wait": "0.0.2",
+    "gulp-webp": "4.0.1",
+    "gulp-zip": "5.0.2",
+    "husky": "4.3.0",
+    "imagemin-jpegoptim": "7.0.0",
+    "stylelint": "13.8.0",
+    "stylelint-config-htmlacademy": "0.1.4",
+    "terser-webpack-plugin": "5.0.3",
+    "webpack": "5.6.0",
+    "webpack-cli": "4.2.0",
+    "webpack-stream": "6.1.1"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.10.4",
-    "normalize.css": "^8.0.1",
-    "picturefill": "^3.0.3",
-    "svg4everybody": "^2.1.9"
+    "normalize.css": "8.0.1"
   },
   "husky": {
     "hooks": {
