@@ -1,21 +1,20 @@
-'use strict';
+import {paths} from '../paths.js';
+import {create as bsCreate} from 'browser-sync';
 
-const { paths: { destination } } = require('../paths');
-const browserSync = require('browser-sync').create('Local Server');
+export const browserSync = bsCreate(`Local Server`);
 
-const server = done => {
+export const server = (done) => {
   browserSync.init({
-    server: destination.root,
+    server: paths.destination.root,
     cors: true,
     open: true,
     notify: false,
     ui: false,
     reloadOnRestart: true,
+    port: 1337,
     // online: false, // Work offline without internet connection
     // tunnel: 'projectname', // Demonstration page: http://projectname.localtunnel.me
   });
 
   done();
 };
-
-module.exports = server;
